Added check if the subscription already excists, plus type fix

+++ modules/communication/inc/serialmanager.h
@@ -28,10 +28,10 @@ circular_buffer_t *SerialManager_GetCircularBufferFromPort(External_Port port);
 void SerialManager_StartTransmission(External_Port port);
 bool SubscribeToCommandI(KFly_Command command,
                          External_Port port,
-                         uint16_t delay_ms);
+                         uint32_t delay_ms);
 bool SubscribeToCommand(KFly_Command command,
                         External_Port port,
-                        uint16_t delay_ms);
+                        uint32_t delay_ms);
 bool UnsubscribeFromCommandI(KFly_Command command, External_Port port);
 bool UnsubscribeFromCommand(KFly_Command command, External_Port port);
 void UnsubscribeFromAllI(void);
